touch ups and test renaming
